Add paths to an object of class 'tidy_sem'

Add paths to an object of class tidy_sem, or replace existing paths. The paths must be specified as model.syntax, and separated by commas.

add_paths(model, ..., strict_check = TRUE)

Arguments

model
An object of class tidy_sem.

...
Paths to add or substitute, specified in lavaan{model.syntax}, and separated by commas.

strict_check
Logical, indicating whether or not to throw an error if the variable names referenced in the new paths do not exist in the data, or in the dictionary, or in the existing syntax element. If strict_check = FALSE, the check is still performed, but throws a warning instead of an error.

Value

An object of class tidy_sem.

Details

Currently, only the lavaan{lavaan} commands ~, ~~, =~, and ~1 are parsed.

See also

Examples

library(lavaan)
df <- iris[, 1:4]
names(df) <- paste0("x_", 1:4)
model <- tidy_sem(df)
model <- measurement(model)
model <- add_paths(model, x =~ a*x_1 + b*x_2 + a*x_3 + b*x_4)
res <- estimate_lavaan(model)

#> Warning: lavaan WARNING: some estimated ov variances are negative

summary(res)

#> lavaan 0.6-5 ended normally after 30 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         12
#>                                                       
#>   Number of observations                           150
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                59.593
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#>   Standard errors                             Standard
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   x =~                                                
#>     x_1        (a)    0.668    0.055   12.159    0.000
#>     x_2        (b)   -0.233    0.030   -7.664    0.000
#>     x_3        (a)    1.853    0.096   19.213    0.000
#>     x_4        (b)    0.686    0.048   14.339    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .x_1               5.843    0.067   86.715    0.000
#>    .x_2               3.057    0.035   86.196    0.000
#>    .x_3               3.758    0.144   26.160    0.000
#>    .x_4               1.199    0.062   19.335    0.000
#>     x                 0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .x_1               0.235    0.026    9.102    0.000
#>    .x_2               0.134    0.014    9.309    0.000
#>    .x_3              -0.339    0.056   -6.041    0.000
#>    .x_4               0.107    0.013    8.300    0.000
#>     x                 1.000                           
#>