Simulated MAC data

This simulated dataset, based on Curry et al., 2019, contains data on moral relevance and judgment across the seven domains of the Morality As Cooperation scale.

Usage

data(curry_mac)

Format

A data.frame with 1392 rows and 42 variables.

Details

sex	factor	Self-identified sex of participants, Male, Female, or Transgendered.

age_years	numeric	Participants' age in years.

KinshipR	numeric	Mean score of moral relevance, kinship subscale.

MutualismR	numeric	Mean score of moral relevance, mutualism subscale.

ExchangeR	numeric	Mean score of moral relevance, exchange subscale.

HawkR	numeric	Mean score of moral relevance, hawk subscale.

DoveR	numeric	Mean score of moral relevance, dove subscale.

DivisionR	numeric	Mean score of moral relevance, division subscale.

PossessionR	numeric	Mean score of moral relevance, possession subscale.

KinshipJ	numeric	Mean score of moral judgment, kinship subscale.

MutualismJ	numeric	Mean score of moral judgment, mutualism subscale.

ExchangeJ	numeric	Mean score of moral judgment, exchange subscale.

HawkJ	numeric	Mean score of moral judgment, hawk subscale.

DoveJ	numeric	Mean score of moral judgment, dove subscale.

DivisionJ	numeric	Mean score of moral judgment, division subscale.

PossessionJ	numeric	Mean score of moral judgment, possession subscale.

References

Curry, O. S., Jones Chesters, M., & Van Lissa, C. J. (2019). Mapping morality with a compass: Testing the theory of 'morality-as-cooperation' with a new questionnaire. Journal of Research in Personality, 78, 106-124. doi:10.1016/j.jrp.2018.10.008