Make a data dictionary

Make a data dictionary for a data.frame, based on variable names.

dictionary(x, split = "_")

Arguments

x
An object for which a method exists, e.g., a vector of variable names, or a data.frame.

split
Character. Within variable names, which character separates the name of the scale/construct from the number (or name) of the item? E.g., if the name is "construct_1", the split character "_" separates the "construct" name from item number "1".

Value

An object of class "data_dict"

Examples

dictionary(c("bfi_1", "bfi_2", "bfi_3", "bfi_4", "bfi_5",
"macqj_1", "macqj_2", "macqj_3", "macqj_4", "macqj_5", "macqj_6",
"macqj_7", "macqj_8", "macqj_9", "macqj_10", "macqj_11",
"macqj_12", "macqj_13", "macqj_14", "macqj_15", "macqj_16",
"macqj_17", "macqj_18", "macqj_19", "macqj_20", "macqj_21",
"macqr_1", "macqr_2", "macqr_3", "macqr_4", "macqr_5", "macqr_6",
"macqr_7", "macqr_8", "macqr_9", "macqr_10", "macqr_11",
"macqr_12", "macqr_13", "macqr_14", "macqr_15", "macqr_16",
"macqr_17", "macqr_18", "macqr_19", "macqr_20", "macqr_21", "sex"))

#>        name scale item     type    label
#> 1     bfi_1   bfi    1 observed    bfi_1
#> 2     bfi_2   bfi    2 observed    bfi_2
#> 3     bfi_3   bfi    3 observed    bfi_3
#> 4     bfi_4   bfi    4 observed    bfi_4
#> 5     bfi_5   bfi    5 observed    bfi_5
#> 6   macqj_1 macqj    1 observed  macqj_1
#> 7   macqj_2 macqj    2 observed  macqj_2
#> 8   macqj_3 macqj    3 observed  macqj_3
#> 9   macqj_4 macqj    4 observed  macqj_4
#> 10  macqj_5 macqj    5 observed  macqj_5
#> 11  macqj_6 macqj    6 observed  macqj_6
#> 12  macqj_7 macqj    7 observed  macqj_7
#> 13  macqj_8 macqj    8 observed  macqj_8
#> 14  macqj_9 macqj    9 observed  macqj_9
#> 15 macqj_10 macqj   10 observed macqj_10
#> 16 macqj_11 macqj   11 observed macqj_11
#> 17 macqj_12 macqj   12 observed macqj_12
#> 18 macqj_13 macqj   13 observed macqj_13
#> 19 macqj_14 macqj   14 observed macqj_14
#> 20 macqj_15 macqj   15 observed macqj_15
#> 21 macqj_16 macqj   16 observed macqj_16
#> 22 macqj_17 macqj   17 observed macqj_17
#> 23 macqj_18 macqj   18 observed macqj_18
#> 24 macqj_19 macqj   19 observed macqj_19
#> 25 macqj_20 macqj   20 observed macqj_20
#> 26 macqj_21 macqj   21 observed macqj_21
#> 27  macqr_1 macqr    1 observed  macqr_1
#> 28  macqr_2 macqr    2 observed  macqr_2
#> 29  macqr_3 macqr    3 observed  macqr_3
#> 30  macqr_4 macqr    4 observed  macqr_4
#> 31  macqr_5 macqr    5 observed  macqr_5
#> 32  macqr_6 macqr    6 observed  macqr_6
#> 33  macqr_7 macqr    7 observed  macqr_7
#> 34  macqr_8 macqr    8 observed  macqr_8
#> 35  macqr_9 macqr    9 observed  macqr_9
#> 36 macqr_10 macqr   10 observed macqr_10
#> 37 macqr_11 macqr   11 observed macqr_11
#> 38 macqr_12 macqr   12 observed macqr_12
#> 39 macqr_13 macqr   13 observed macqr_13
#> 40 macqr_14 macqr   14 observed macqr_14
#> 41 macqr_15 macqr   15 observed macqr_15
#> 42 macqr_16 macqr   16 observed macqr_16
#> 43 macqr_17 macqr   17 observed macqr_17
#> 44 macqr_18 macqr   18 observed macqr_18
#> 45 macqr_19 macqr   19 observed macqr_19
#> 46 macqr_20 macqr   20 observed macqr_20
#> 47 macqr_21 macqr   21 observed macqr_21
#> 48      sex  <NA> <NA> observed      sex

dictionary(c("bfi_1", "bfi_2", "bfi_3", "bfi_4", "bfi_5",
"mac_q_j_1", "mac_q_j_2", "mac_q_j_3", "mac_q_j_4", "mac_q_j_5", "mac_q_j_6",
"mac_q_j_7", "mac_q_j_8", "mac_q_j_9", "mac_q_j_10", "mac_q_j_11",
"mac_q_j_12", "mac_q_j_13", "mac_q_j_14", "mac_q_j_15", "mac_q_j_16",
"mac_q_j_17", "mac_q_j_18", "mac_q_j_19", "mac_q_j_20", "mac_q_j_21",
"mac_q_r_1", "mac_q_r_2", "mac_q_r_3", "mac_q_r_4", "mac_q_r_5", "mac_q_r_6",
"mac_q_r_7", "mac_q_r_8", "mac_q_r_9", "mac_q_r_10", "mac_q_r_11",
"mac_q_r_12", "mac_q_r_13", "mac_q_r_14", "mac_q_r_15", "mac_q_r_16",
"mac_q_r_17", "mac_q_r_18", "mac_q_r_19", "mac_q_r_20", "mac_q_r_21"))

#> Warning: Some variable names contained more than one splitting character. Ignored all but the last splitting character.

#>          name   scale item     type      label
#> 1       bfi_1     bfi    1 observed      bfi_1
#> 2       bfi_2     bfi    2 observed      bfi_2
#> 3       bfi_3     bfi    3 observed      bfi_3
#> 4       bfi_4     bfi    4 observed      bfi_4
#> 5       bfi_5     bfi    5 observed      bfi_5
#> 6   mac_q_j_1 mac_q_j    1 observed  mac_q_j_1
#> 7   mac_q_j_2 mac_q_j    2 observed  mac_q_j_2
#> 8   mac_q_j_3 mac_q_j    3 observed  mac_q_j_3
#> 9   mac_q_j_4 mac_q_j    4 observed  mac_q_j_4
#> 10  mac_q_j_5 mac_q_j    5 observed  mac_q_j_5
#> 11  mac_q_j_6 mac_q_j    6 observed  mac_q_j_6
#> 12  mac_q_j_7 mac_q_j    7 observed  mac_q_j_7
#> 13  mac_q_j_8 mac_q_j    8 observed  mac_q_j_8
#> 14  mac_q_j_9 mac_q_j    9 observed  mac_q_j_9
#> 15 mac_q_j_10 mac_q_j   10 observed mac_q_j_10
#> 16 mac_q_j_11 mac_q_j   11 observed mac_q_j_11
#> 17 mac_q_j_12 mac_q_j   12 observed mac_q_j_12
#> 18 mac_q_j_13 mac_q_j   13 observed mac_q_j_13
#> 19 mac_q_j_14 mac_q_j   14 observed mac_q_j_14
#> 20 mac_q_j_15 mac_q_j   15 observed mac_q_j_15
#> 21 mac_q_j_16 mac_q_j   16 observed mac_q_j_16
#> 22 mac_q_j_17 mac_q_j   17 observed mac_q_j_17
#> 23 mac_q_j_18 mac_q_j   18 observed mac_q_j_18
#> 24 mac_q_j_19 mac_q_j   19 observed mac_q_j_19
#> 25 mac_q_j_20 mac_q_j   20 observed mac_q_j_20
#> 26 mac_q_j_21 mac_q_j   21 observed mac_q_j_21
#> 27  mac_q_r_1 mac_q_r    1 observed  mac_q_r_1
#> 28  mac_q_r_2 mac_q_r    2 observed  mac_q_r_2
#> 29  mac_q_r_3 mac_q_r    3 observed  mac_q_r_3
#> 30  mac_q_r_4 mac_q_r    4 observed  mac_q_r_4
#> 31  mac_q_r_5 mac_q_r    5 observed  mac_q_r_5
#> 32  mac_q_r_6 mac_q_r    6 observed  mac_q_r_6
#> 33  mac_q_r_7 mac_q_r    7 observed  mac_q_r_7
#> 34  mac_q_r_8 mac_q_r    8 observed  mac_q_r_8
#> 35  mac_q_r_9 mac_q_r    9 observed  mac_q_r_9
#> 36 mac_q_r_10 mac_q_r   10 observed mac_q_r_10
#> 37 mac_q_r_11 mac_q_r   11 observed mac_q_r_11
#> 38 mac_q_r_12 mac_q_r   12 observed mac_q_r_12
#> 39 mac_q_r_13 mac_q_r   13 observed mac_q_r_13
#> 40 mac_q_r_14 mac_q_r   14 observed mac_q_r_14
#> 41 mac_q_r_15 mac_q_r   15 observed mac_q_r_15
#> 42 mac_q_r_16 mac_q_r   16 observed mac_q_r_16
#> 43 mac_q_r_17 mac_q_r   17 observed mac_q_r_17
#> 44 mac_q_r_18 mac_q_r   18 observed mac_q_r_18
#> 45 mac_q_r_19 mac_q_r   19 observed mac_q_r_19
#> 46 mac_q_r_20 mac_q_r   20 observed mac_q_r_20
#> 47 mac_q_r_21 mac_q_r   21 observed mac_q_r_21