Simulated empathy data

This simulated dataset, based on Van Lissa et al., 2014, contains six annual assessments of adolescents' mean scores on the empathic concern and perspective taking subscales of the Interpersonal Reactivity Index (Davis, 1983). The first measurement wave occurred when adolescents were, on average, 13 years old, and the last one when they were 18 years old.

Usage

data(empathy)

Format

A data frame with 467 rows and 13 variables.

Details

ec1	numeric	Mean score of empathic concern in wave 1

ec2	numeric	Mean score of empathic concern in wave 2

ec3	numeric	Mean score of empathic concern in wave 3

ec4	numeric	Mean score of empathic concern in wave 4

ec5	numeric	Mean score of empathic concern in wave 5

ec6	numeric	Mean score of empathic concern in wave 6

pt1	numeric	Mean score of perspective taking in wave 1

pt2	numeric	Mean score of perspective taking in wave 2

pt3	numeric	Mean score of perspective taking in wave 3

pt4	numeric	Mean score of perspective taking in wave 4

pt5	numeric	Mean score of perspective taking in wave 5

pt6	numeric	Mean score of perspective taking in wave 6

sex	factor	Adolescent sex; M = male, F = female.

References

Van Lissa, C. J., Hawk, S. T., Branje, S. J., Koot, H. M., Van Lier, P. A., & Meeus, W. H. (2014). Divergence Between Adolescent and Parental Perceptions of Conflict in Relationship to Adolescent Empathy Development. Journal of Youth and Adolescence, (Journal Article), 1–14. doi:10.1007/s10964-014-0152-5