Estimate tidy_sem using 'lavaan'

This function is a wrapper for the lavaan estimating functions. By default, the wrapper uses sem, but users can also specify lavaan, cfa, or growth.

estimate_lavaan(x, func = "sem", ...)

Arguments

x
An object of class tidy_sem.

func
The lavaan modeling function to invoke, Default: 'sem'.

...
Additional parameters passed to the estimating function.

Value

An object of class lavaan.

Examples

library(lavaan)
model <- tidy_sem(iris, "\\.")
model <- measurement(model, center = F, scale = F)
res <- estimate_lavaan(model)

#> Warning: lavaan WARNING: some estimated ov variances are negative

summary(res)

#> lavaan 0.6-5 ended normally after 58 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                           150
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                               265.801
#>   Degrees of freedom                                 3
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#>   Standard errors                             Standard
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   Sepal =~                                            
#>     Sepal.Length      1.000                           
#>     Sepal.Width       0.510    0.009   57.074    0.000
#>   Petal =~                                            
#>     Petal.Length      1.000                           
#>     Petal.Width       0.330    0.005   63.864    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   Sepal ~~                                            
#>     Petal             1.251    0.154    8.138    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .Sepal.Length      0.000                           
#>    .Sepal.Width       0.000                           
#>     Sepal             5.831    0.074   78.678    0.000
#>    .Petal.Length      0.000                           
#>    .Petal.Width       0.000                           
#>     Petal             3.826    0.131   29.264    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .Sepal.Length     -0.128    0.045   -2.863    0.004
#>    .Sepal.Width       0.448    0.053    8.472    0.000
#>     Sepal             0.952    0.107    8.894    0.000
#>    .Petal.Length     -0.324    0.060   -5.385    0.000
#>    .Petal.Width       0.102    0.013    7.934    0.000
#>     Petal             2.907    0.299    9.717    0.000
#>