Make a data dictionary

Make a data dictionary for a data.frame, based on variable names.

get_dictionary(x, split = "_")

Arguments

x
An object for which a method exists, e.g., a vector of variable names, or a data.frame.

split
Character. Within variable names, which character separates the name of the scale/construct from the number (or name) of the item? E.g., if the name is "construct_1", the split character "_" separates the "construct" name from item number "1".

Value

An object of class "data_dict"

Examples

get_dictionary(c("bfi_1", "bfi_2", "bfi_3", "bfi_4", "bfi_5",
"macqj_1", "macqj_2", "macqj_3", "macqj_4", "macqj_5", "macqj_6",
"macqj_7", "macqj_8", "macqj_9", "macqj_10", "macqj_11",
"macqj_12", "macqj_13", "macqj_14", "macqj_15", "macqj_16",
"macqj_17", "macqj_18", "macqj_19", "macqj_20", "macqj_21",
"macqr_1", "macqr_2", "macqr_3", "macqr_4", "macqr_5", "macqr_6",
"macqr_7", "macqr_8", "macqr_9", "macqr_10", "macqr_11",
"macqr_12", "macqr_13", "macqr_14", "macqr_15", "macqr_16",
"macqr_17", "macqr_18", "macqr_19", "macqr_20", "macqr_21", "sex"))

#>        name scale     type    label
#> 1     bfi_1   bfi observed    bfi_1
#> 2     bfi_2   bfi observed    bfi_2
#> 3     bfi_3   bfi observed    bfi_3
#> 4     bfi_4   bfi observed    bfi_4
#> 5     bfi_5   bfi observed    bfi_5
#> 6   macqj_1 macqj observed  macqj_1
#> 7   macqj_2 macqj observed  macqj_2
#> 8   macqj_3 macqj observed  macqj_3
#> 9   macqj_4 macqj observed  macqj_4
#> 10  macqj_5 macqj observed  macqj_5
#> 11  macqj_6 macqj observed  macqj_6
#> 12  macqj_7 macqj observed  macqj_7
#> 13  macqj_8 macqj observed  macqj_8
#> 14  macqj_9 macqj observed  macqj_9
#> 15 macqj_10 macqj observed macqj_10
#> 16 macqj_11 macqj observed macqj_11
#> 17 macqj_12 macqj observed macqj_12
#> 18 macqj_13 macqj observed macqj_13
#> 19 macqj_14 macqj observed macqj_14
#> 20 macqj_15 macqj observed macqj_15
#> 21 macqj_16 macqj observed macqj_16
#> 22 macqj_17 macqj observed macqj_17
#> 23 macqj_18 macqj observed macqj_18
#> 24 macqj_19 macqj observed macqj_19
#> 25 macqj_20 macqj observed macqj_20
#> 26 macqj_21 macqj observed macqj_21
#> 27  macqr_1 macqr observed  macqr_1
#> 28  macqr_2 macqr observed  macqr_2
#> 29  macqr_3 macqr observed  macqr_3
#> 30  macqr_4 macqr observed  macqr_4
#> 31  macqr_5 macqr observed  macqr_5
#> 32  macqr_6 macqr observed  macqr_6
#> 33  macqr_7 macqr observed  macqr_7
#> 34  macqr_8 macqr observed  macqr_8
#> 35  macqr_9 macqr observed  macqr_9
#> 36 macqr_10 macqr observed macqr_10
#> 37 macqr_11 macqr observed macqr_11
#> 38 macqr_12 macqr observed macqr_12
#> 39 macqr_13 macqr observed macqr_13
#> 40 macqr_14 macqr observed macqr_14
#> 41 macqr_15 macqr observed macqr_15
#> 42 macqr_16 macqr observed macqr_16
#> 43 macqr_17 macqr observed macqr_17
#> 44 macqr_18 macqr observed macqr_18
#> 45 macqr_19 macqr observed macqr_19
#> 46 macqr_20 macqr observed macqr_20
#> 47 macqr_21 macqr observed macqr_21
#> 48      sex  <NA> observed      sex

get_dictionary(c("bfi_1", "bfi_2", "bfi_3", "bfi_4", "bfi_5",
"mac_q_j_1", "mac_q_j_2", "mac_q_j_3", "mac_q_j_4", "mac_q_j_5", "mac_q_j_6",
"mac_q_j_7", "mac_q_j_8", "mac_q_j_9", "mac_q_j_10", "mac_q_j_11",
"mac_q_j_12", "mac_q_j_13", "mac_q_j_14", "mac_q_j_15", "mac_q_j_16",
"mac_q_j_17", "mac_q_j_18", "mac_q_j_19", "mac_q_j_20", "mac_q_j_21",
"mac_q_r_1", "mac_q_r_2", "mac_q_r_3", "mac_q_r_4", "mac_q_r_5", "mac_q_r_6",
"mac_q_r_7", "mac_q_r_8", "mac_q_r_9", "mac_q_r_10", "mac_q_r_11",
"mac_q_r_12", "mac_q_r_13", "mac_q_r_14", "mac_q_r_15", "mac_q_r_16",
"mac_q_r_17", "mac_q_r_18", "mac_q_r_19", "mac_q_r_20", "mac_q_r_21"))

#> Some variable names contained more than one splitting character. Ignored all but the last splitting character.FALSE

#>          name   scale     type      label
#> 1       bfi_1     bfi observed      bfi_1
#> 2       bfi_2     bfi observed      bfi_2
#> 3       bfi_3     bfi observed      bfi_3
#> 4       bfi_4     bfi observed      bfi_4
#> 5       bfi_5     bfi observed      bfi_5
#> 6   mac_q_j_1 mac_q_j observed  mac_q_j_1
#> 7   mac_q_j_2 mac_q_j observed  mac_q_j_2
#> 8   mac_q_j_3 mac_q_j observed  mac_q_j_3
#> 9   mac_q_j_4 mac_q_j observed  mac_q_j_4
#> 10  mac_q_j_5 mac_q_j observed  mac_q_j_5
#> 11  mac_q_j_6 mac_q_j observed  mac_q_j_6
#> 12  mac_q_j_7 mac_q_j observed  mac_q_j_7
#> 13  mac_q_j_8 mac_q_j observed  mac_q_j_8
#> 14  mac_q_j_9 mac_q_j observed  mac_q_j_9
#> 15 mac_q_j_10 mac_q_j observed mac_q_j_10
#> 16 mac_q_j_11 mac_q_j observed mac_q_j_11
#> 17 mac_q_j_12 mac_q_j observed mac_q_j_12
#> 18 mac_q_j_13 mac_q_j observed mac_q_j_13
#> 19 mac_q_j_14 mac_q_j observed mac_q_j_14
#> 20 mac_q_j_15 mac_q_j observed mac_q_j_15
#> 21 mac_q_j_16 mac_q_j observed mac_q_j_16
#> 22 mac_q_j_17 mac_q_j observed mac_q_j_17
#> 23 mac_q_j_18 mac_q_j observed mac_q_j_18
#> 24 mac_q_j_19 mac_q_j observed mac_q_j_19
#> 25 mac_q_j_20 mac_q_j observed mac_q_j_20
#> 26 mac_q_j_21 mac_q_j observed mac_q_j_21
#> 27  mac_q_r_1 mac_q_r observed  mac_q_r_1
#> 28  mac_q_r_2 mac_q_r observed  mac_q_r_2
#> 29  mac_q_r_3 mac_q_r observed  mac_q_r_3
#> 30  mac_q_r_4 mac_q_r observed  mac_q_r_4
#> 31  mac_q_r_5 mac_q_r observed  mac_q_r_5
#> 32  mac_q_r_6 mac_q_r observed  mac_q_r_6
#> 33  mac_q_r_7 mac_q_r observed  mac_q_r_7
#> 34  mac_q_r_8 mac_q_r observed  mac_q_r_8
#> 35  mac_q_r_9 mac_q_r observed  mac_q_r_9
#> 36 mac_q_r_10 mac_q_r observed mac_q_r_10
#> 37 mac_q_r_11 mac_q_r observed mac_q_r_11
#> 38 mac_q_r_12 mac_q_r observed mac_q_r_12
#> 39 mac_q_r_13 mac_q_r observed mac_q_r_13
#> 40 mac_q_r_14 mac_q_r observed mac_q_r_14
#> 41 mac_q_r_15 mac_q_r observed mac_q_r_15
#> 42 mac_q_r_16 mac_q_r observed mac_q_r_16
#> 43 mac_q_r_17 mac_q_r observed mac_q_r_17
#> 44 mac_q_r_18 mac_q_r observed mac_q_r_18
#> 45 mac_q_r_19 mac_q_r observed mac_q_r_19
#> 46 mac_q_r_20 mac_q_r observed mac_q_r_20
#> 47 mac_q_r_21 mac_q_r observed mac_q_r_21