Estimate an Auxiliary Model using the BCH Method

Estimate an auxiliary model based on a latent classification by means of mixture modeling (see mx_mixture).

The auxiliary model is treated as a multi-group model. All cases are used in all groups, but they are weighted by group-specific BCH weights as described in Bolck, Croon, & Hagenaars, 2004.

BCH(x, model, data, ...)

Arguments

x
An object for which a method exists.

model
An object that can be converted to an OpenMx model using as_ram.

data
A data.frame on which the auxiliary model can be evaluated.

...
further arguments to be passed to or from other methods.

Value

An MxModel.

References

Bolck, A., Croon, M., & Hagenaars, J. (2004). Estimating latent structure models with categorical variables: One-step versus three-step estimators. Political Analysis, 12(1), 3–27. <doi:10.2307/25791751>

Examples

dat <- data.frame(x = iris$Petal.Length)
mixmod <- mx_profiles(dat,
                      classes = 2)

#> Called from: tidySEM:::mixture_starts(model = new("MxModel", name = "mix2", 
#>     matrices = list(weights = new("FullMatrix", name = "weights", 
#>         values = c(1, 1), labels = c(NA_character_, NA_character_
#>         ), free = c(FALSE, TRUE), lbound = c(1e-04, 1e-04), ubound = c(NA_real_, 
#>         NA_real_), .squareBrackets = numeric(0), .persist = TRUE, 
#>         .condenseSlots = FALSE, display = character(0), dependencies = integer(0), 
#>         joinModel = NA_character_, joinKey = NA_character_)), 
#>     algebras = list(), constraints = list(), intervals = list(), 
#>     latentVars = character(0), manifestVars = character(0), data = new("MxDataStatic", 
#>         observed = list(x = c(1.4, 1.4, 1.3, 1.5, 1.4, 1.7, 1.4, 
#>         1.5, 1.4, 1.5, 1.5, 1.6, 1.4, 1.1, 1.2, 1.5, 1.3, 1.4, 
#>         1.7, 1.5, 1.7, 1.5, 1, 1.7, 1.9, 1.6, 1.6, 1.5, 1.4, 
#>         1.6, 1.6, 1.5, 1.5, 1.4, 1.5, 1.2, 1.3, 1.4, 1.3, 1.5, 
#>         1.3, 1.3, 1.3, 1.6, 1.9, 1.4, 1.6, 1.4, 1.5, 1.4, 4.7, 
#>         4.5, 4.9, 4, 4.6, 4.5, 4.7, 3.3, 4.6, 3.9, 3.5, 4.2, 
#>         4, 4.7, 3.6, 4.4, 4.5, 4.1, 4.5, 3.9, 4.8, 4, 4.9, 4.7, 
#>         4.3, 4.4, 4.8, 5, 4.5, 3.5, 3.8, 3.7, 3.9, 5.1, 4.5, 
#>         4.5, 4.7, 4.4, 4.1, 4, 4.4, 4.6, 4, 3.3, 4.2, 4.2, 4.2, 
#>         4.3, 3, 4.1, 6, 5.1, 5.9, 5.6, 5.8, 6.6, 4.5, 6.3, 5.8, 
#>         6.1, 5.1, 5.3, 5.5, 5, 5.1, 5.3, 5.5, 6.7, 6.9, 5, 5.7, 
#>         4.9, 6.7, 4.9, 5.7, 6, 4.8, 4.9, 5.6, 5.8, 6.1, 6.4, 
#>         5.6, 5.1, 5.6, 6.1, 5.6, 5.5, 4.8, 5.4, 5.6, 5.1, 5.1, 
#>         5.9, 5.7, 5.2, 5, 5.2, 5.4, 5.1)), means = NA_real_, 
#>         type = "raw", numObs = 150, observedStats = list(), .isSorted = FALSE, 
#>         .needSort = TRUE, .parallel = TRUE, .noExoOptimize = TRUE, 
#>         primaryKey = NA_character_, weight = NA_character_, frequency = NA_character_, 
#>         minVariance = 1.49011611938477e-08, algebra = character(0), 
#>         warnNPDuseWeight = TRUE, exoFree = NULL, naAction = "pass", 
#>         verbose = 0L, name = "data"), submodels = list(class1 = new("MxRAMModel", 
#>         name = "class1", matrices = list(A = new("FullMatrix", 
#>             name = "A", values = 0, labels = NA_character_, free = FALSE, 
#>             lbound = NA_real_, ubound = NA_real_, .squareBrackets = numeric(0), 
#>             .persist = TRUE, .condenseSlots = FALSE, display = character(0), 
#>             dependencies = integer(0), joinModel = NA_character_, 
#>             joinKey = NA_character_), S = new("SymmMatrix", name = "S", 
#>             values = 0, labels = "v1", free = TRUE, lbound = NA_real_, 
#>             ubound = NA_real_, .squareBrackets = numeric(0), 
#>             .persist = TRUE, .condenseSlots = FALSE, display = character(0), 
#>             dependencies = integer(0), joinModel = NA_character_, 
#>             joinKey = NA_character_), F = new("FullMatrix", name = "F", 
#>             values = 1, labels = NA_character_, free = FALSE, 
#>             lbound = NA_real_, ubound = NA_real_, .squareBrackets = numeric(0), 
#>             .persist = TRUE, .condenseSlots = TRUE, display = character(0), 
#>             dependencies = integer(0), joinModel = NA_character_, 
#>             joinKey = NA_character_), M = new("FullMatrix", name = "M", 
#>             values = 0, labels = "m11", free = TRUE, lbound = NA_real_, 
#>             ubound = NA_real_, .squareBrackets = numeric(0), 
#>             .persist = TRUE, .condenseSlots = FALSE, display = character(0), 
#>             dependencies = integer(0), joinModel = NA_character_, 
#>             joinKey = NA_character_)), algebras = list(), constraints = list(), 
#>         intervals = list(), latentVars = character(0), manifestVars = "x", 
#>         data = NULL, submodels = list(), expectation = new("MxExpectationRAM", 
#>             A = "A", S = "S", F = "F", M = "M", thresholds = NA_character_, 
#>             dims = NA_character_, usePPML = FALSE, ppmlData = NULL, 
#>             UnfilteredExpCov = NA, numStats = numeric(0), between = NULL, 
#>             isProductNode = FALSE, verbose = 0L, .rampartCycleLimit = NA_integer_, 
#>             .rampartUnitLimit = NA_integer_, .useSufficientSets = TRUE, 
#>             .forceSingleGroup = FALSE, .analyzeDefVars = TRUE, 
#>             .maxDebugGroups = 0L, .optimizeMean = 2L, .useSparse = NA, 
#>             expectedCovariance = NULL, expectedMean = NULL, expectedFullCovariance = NULL, 
#>             expectedFullMean = NULL, threshnames = NA_character_, 
#>             discrete = NA_character_, discreteSpec = NULL, .discreteCheckCount = TRUE, 
#>             selectionPlan = NULL, selectionVector = NA_character_, 
#>             data = NA_integer_, dataColumns = integer(0), dataColumnNames = NULL, 
#>             .runDims = character(0), output = list(), debug = list(), 
#>             name = "expectation"), fitfunction = new("MxFitFunctionML", 
#>             fellner = NA, verbose = 0L, profileOut = character(0), 
#>             rowwiseParallel = NA, jointConditionOn = "auto", 
#>             components = character(0), info = list(), dependencies = integer(0), 
#>             expectation = integer(0), vector = TRUE, rowDiagnostics = FALSE, 
#>             result = numeric(0), name = "fitfunction"), compute = NULL, 
#>         independent = FALSE, options = list(), output = list(), 
#>         runstate = list(), .newobjects = FALSE, .resetdata = FALSE, 
#>         .wasRun = FALSE, .modifiedSinceRun = TRUE, .version = "2.19.6"), 
#>         class2 = new("MxRAMModel", name = "class2", matrices = list(
#>             A = new("FullMatrix", name = "A", values = 0, labels = NA_character_, 
#>                 free = FALSE, lbound = NA_real_, ubound = NA_real_, 
#>                 .squareBrackets = numeric(0), .persist = TRUE, 
#>                 .condenseSlots = FALSE, display = character(0), 
#>                 dependencies = integer(0), joinModel = NA_character_, 
#>                 joinKey = NA_character_), S = new("SymmMatrix", 
#>                 name = "S", values = 0, labels = "v1", free = TRUE, 
#>                 lbound = NA_real_, ubound = NA_real_, .squareBrackets = numeric(0), 
#>                 .persist = TRUE, .condenseSlots = FALSE, display = character(0), 
#>                 dependencies = integer(0), joinModel = NA_character_, 
#>                 joinKey = NA_character_), F = new("FullMatrix", 
#>                 name = "F", values = 1, labels = NA_character_, 
#>                 free = FALSE, lbound = NA_real_, ubound = NA_real_, 
#>                 .squareBrackets = numeric(0), .persist = TRUE, 
#>                 .condenseSlots = TRUE, display = character(0), 
#>                 dependencies = integer(0), joinModel = NA_character_, 
#>                 joinKey = NA_character_), M = new("FullMatrix", 
#>                 name = "M", values = 0, labels = "m21", free = TRUE, 
#>                 lbound = NA_real_, ubound = NA_real_, .squareBrackets = numeric(0), 
#>                 .persist = TRUE, .condenseSlots = FALSE, display = character(0), 
#>                 dependencies = integer(0), joinModel = NA_character_, 
#>                 joinKey = NA_character_)), algebras = list(), 
#>             constraints = list(), intervals = list(), latentVars = character(0), 
#>             manifestVars = "x", data = NULL, submodels = list(), 
#>             expectation = new("MxExpectationRAM", A = "A", S = "S", 
#>                 F = "F", M = "M", thresholds = NA_character_, 
#>                 dims = NA_character_, usePPML = FALSE, ppmlData = NULL, 
#>                 UnfilteredExpCov = NA, numStats = numeric(0), 
#>                 between = NULL, isProductNode = FALSE, verbose = 0L, 
#>                 .rampartCycleLimit = NA_integer_, .rampartUnitLimit = NA_integer_, 
#>                 .useSufficientSets = TRUE, .forceSingleGroup = FALSE, 
#>                 .analyzeDefVars = TRUE, .maxDebugGroups = 0L, 
#>                 .optimizeMean = 2L, .useSparse = NA, expectedCovariance = NULL, 
#>                 expectedMean = NULL, expectedFullCovariance = NULL, 
#>                 expectedFullMean = NULL, threshnames = NA_character_, 
#>                 discrete = NA_character_, discreteSpec = NULL, 
#>                 .discreteCheckCount = TRUE, selectionPlan = NULL, 
#>                 selectionVector = NA_character_, data = NA_integer_, 
#>                 dataColumns = integer(0), dataColumnNames = NULL, 
#>                 .runDims = character(0), output = list(), debug = list(), 
#>                 name = "expectation"), fitfunction = new("MxFitFunctionML", 
#>                 fellner = NA, verbose = 0L, profileOut = character(0), 
#>                 rowwiseParallel = NA, jointConditionOn = "auto", 
#>                 components = character(0), info = list(), dependencies = integer(0), 
#>                 expectation = integer(0), vector = TRUE, rowDiagnostics = FALSE, 
#>                 result = numeric(0), name = "fitfunction"), compute = NULL, 
#>             independent = FALSE, options = list(), output = list(), 
#>             runstate = list(), .newobjects = FALSE, .resetdata = FALSE, 
#>             .wasRun = FALSE, .modifiedSinceRun = TRUE, .version = "2.19.6")), 
#>     expectation = new("MxExpectationMixture", weights = "weights", 
#>         components = c("class1", "class2"), verbose = 0L, scale = "softmax", 
#>         data = NA_integer_, dataColumns = integer(0), dataColumnNames = NULL, 
#>         .runDims = character(0), output = list(), debug = list(), 
#>         name = "expectation"), fitfunction = new("MxFitFunctionML", 
#>         fellner = NA, verbose = 0L, profileOut = character(0), 
#>         rowwiseParallel = NA, jointConditionOn = "auto", components = character(0), 
#>         info = list(), dependencies = integer(0), expectation = integer(0), 
#>         vector = FALSE, rowDiagnostics = FALSE, result = numeric(0), 
#>         name = "fitfunction"), compute = NULL, independent = FALSE, 
#>     options = list(), output = list(), runstate = list(), .newobjects = FALSE, 
#>     .resetdata = FALSE, .wasRun = FALSE, .modifiedSinceRun = TRUE, 
#>     .version = "2.19.6"), classes = 2, data = dat)
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#376: stopifnot(`mxModel is not a mixture model.` = inherits(model@expectation, 
#>     "MxExpectationMixture") | attr(model, "tidySEM") == "mixture")
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#377: stopifnot(`mxModel must contain data to determine starting values.` = !(is.null(model@data) | 
#>     is.null(model@data$observed)))
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#378: classes <- length(model@submodels)
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#379: if (classes < 2) {
#>     strts <- try({
#>         mxAutoStart(model, type = "ULS")
#>     })
#>     if (inherits(strts, "try-error")) {
#>         strts <- try({
#>             mxRun(model)
#>         })
#>     }
#>     if (inherits(strts, "try-error")) {
#>         stop("Could not derive suitable starting values for the 1-class model.")
#>     }
#>     else {
#>         return(strts)
#>     }
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#390: data <- model@data$observed
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#391: if (!hasArg(splits)) {
#>     splits <- try({
#>         kmeans(x = data, centers = classes)$cluster
#>     })
#>     if (inherits(splits, "try-error")) {
#>         message("Could not initialize clusters using K-means.")
#>         splits <- try({
#>             cutree(hclust(dist(data)), k = classes)
#>         })
#>         if (inherits(splits, "try-error")) {
#>             stop("Could not initialize clusters using hierarchical clustering. Consider using a different clustering method, or imputing missing data.")
#>         }
#>     }
#> } else {
#>     stopifnot(`Number of unique values in splits must be identical to the number of latent classes.` = length(unique(splits) == 
#>         names(model@submodels)))
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#392: splits <- try({
#>     kmeans(x = data, centers = classes)$cluster
#> })
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#392: kmeans(x = data, centers = classes)$cluster
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#393: if (inherits(splits, "try-error")) {
#>     message("Could not initialize clusters using K-means.")
#>     splits <- try({
#>         cutree(hclust(dist(data)), k = classes)
#>     })
#>     if (inherits(splits, "try-error")) {
#>         stop("Could not initialize clusters using hierarchical clustering. Consider using a different clustering method, or imputing missing data.")
#>     }
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#404: tab_split <- table(splits)
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#405: if (any(tab_split) < 2) {
#>     small_cats <- which(tab_split < 2)
#>     choose_from <- which(tab_split > 2 + length(small_cats))
#>     if (length(choose_from) == 0) 
#>         stop("Some clusters were too small to determine sensible starting values in `mixture_starts()`. Either specify splits manually, or reduce the number of classes.")
#>     splits[sample(which(splits %in% choose_from), length(small_cats))] <- small_cats
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#406: small_cats <- which(tab_split < 2)
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#407: choose_from <- which(tab_split > 2 + length(small_cats))
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#408: if (length(choose_from) == 0) stop("Some clusters were too small to determine sensible starting values in `mixture_starts()`. Either specify splits manually, or reduce the number of classes.")
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#409: splits[sample(which(splits %in% choose_from), length(small_cats))] <- small_cats
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#412: if (!classes == length(unique(splits))) {
#>     stop("Argument 'splits' does not identify a number of groups equal to 'classes'.")
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#415: if (!all(unique(splits) %in% 1:classes)) {
#>     splits <- as.integer(as.factor(splits))
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#419: strts <- lapply(1:classes, function(i) {
#>     thissub <- names(model@submodels)[i]
#>     mxModel(model[[thissub]], mxData(data[splits == i, , drop = FALSE], 
#>         type = "raw"), mxFitFunctionML())
#> })
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#425: strts <- do.call(mxModel, c(list(model = "mg_starts", mxFitFunctionMultigroup(names(model@submodels)), 
#>     strts)))
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#426: strts <- try({
#>     strts <- mxAutoStart(strts, type = "ULS")
#>     mxRun(strts, silent = TRUE, suppressWarnings = TRUE)
#> })
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#427: strts <- mxAutoStart(strts, type = "ULS")
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#428: mxRun(strts, silent = TRUE, suppressWarnings = TRUE)
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#430: if (inherits(strts, "try-error")) {
#>     strts <- try({
#>         strts <- mxAutoStart(strts, type = "DWLS")
#>         strts <<- mxTryHard(strts, extraTries = 100, silent = TRUE, 
#>             verbose = FALSE, bestInitsOutput = FALSE)
#>     })
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#440: if (inherits(strts, "try-error")) {
#>     strts <- try({
#>         mxRun(model)
#>     })
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#443: if (inherits(strts, "try-error")) {
#>     stop("Could not derive suitable starting values for the ", 
#>         classes, "-class model.")
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#447: for (i in names(model@submodels)) {
#>     for (mtx in names(model[[i]]@matrices)) {
#>         model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#>     }
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#448: for (mtx in names(model[[i]]@matrices)) {
#>     model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#449: model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#449: model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#449: model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#449: model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#448: for (mtx in names(model[[i]]@matrices)) {
#>     model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> }
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#449: model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#449: model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#449: model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#449: model[[i]][[mtx]]$values <- strts[[i]][[mtx]]$values
#> debug at C:/Git_Repositories/tidySEM/R/mx_mixture.R#452: return(model)

#> Running mix2 with 4 parameters

#> Running mix2 with 4 parameters

res <- BCH(mixmod, "y ~ 1", data = data.frame(y = iris$Sepal.Length))

#> Running aux with 4 parameters