Append Pseudo-class Draws

Generates m datasets with random draws of a variable named class, with probability for these draws based on each case's probability of belonging to that class according to the model in x.

append_class_draws(x, data = NULL, m = 20)

Arguments

x

An object for which a method exists, usually a mx_mixture model.

data

A data.frame which the class variable is appended to. Note that the row order must be identical to that of the data used to fit x, as these data will be augmented with a pseudo-class draw for that specific individual.

m

Integer. Number of datasets to generate. Default is 10.

Value

A data.frame of class class_draws.

Examples

dat <- iris[c(1:5, 50:55, 100:105),1:3]
colnames(dat) <- letters[1:3]
fit <- mx_profiles(data = dat, classes = 2)
#> Running mix2 with 10 parameters
#> Running mix2 with 10 parameters

append_class_draws(fit, data = iris[c(1:5, 50:55, 100:105), 4, drop = FALSE])
#>       id_dataset Petal.Width class
#> 1              1         0.2     2
#> 2              1         0.2     2
#> 3              1         0.2     2
#> 4              1         0.2     2
#> 5              1         0.2     2
#> 50             1         0.2     2
#> 51             1         1.4     1
#> 52             1         1.5     1
#> 53             1         1.5     1
#> 54             1         1.3     1
#> 55             1         1.5     1
#> 100            1         1.3     1
#> 101            1         2.5     1
#> 102            1         1.9     1
#> 103            1         2.1     1
#> 104            1         1.8     1
#> 105            1         2.2     1
#> 11             2         0.2     2
#> 21             2         0.2     2
#> 31             2         0.2     2
#> 41             2         0.2     2
#> 56             2         0.2     2
#> 501            2         0.2     2
#> 511            2         1.4     1
#> 521            2         1.5     1
#> 531            2         1.5     1
#> 541            2         1.3     1
#> 551            2         1.5     1
#> 1001           2         1.3     1
#> 1011           2         2.5     1
#> 1021           2         1.9     1
#> 1031           2         2.1     1
#> 1041           2         1.8     1
#> 1051           2         2.2     1
#> 12             3         0.2     2
#> 22             3         0.2     2
#> 32             3         0.2     2
#> 42             3         0.2     2
#> 57             3         0.2     2
#> 502            3         0.2     2
#> 512            3         1.4     1
#> 522            3         1.5     1
#> 532            3         1.5     1
#> 542            3         1.3     1
#> 552            3         1.5     1
#> 1002           3         1.3     1
#> 1012           3         2.5     1
#> 1022           3         1.9     1
#> 1032           3         2.1     1
#> 1042           3         1.8     1
#> 1052           3         2.2     1
#> 13             4         0.2     2
#> 23             4         0.2     2
#> 33             4         0.2     2
#> 43             4         0.2     2
#> 58             4         0.2     2
#> 503            4         0.2     2
#> 513            4         1.4     1
#> 523            4         1.5     1
#> 533            4         1.5     1
#> 543            4         1.3     1
#> 553            4         1.5     1
#> 1003           4         1.3     1
#> 1013           4         2.5     1
#> 1023           4         1.9     1
#> 1033           4         2.1     1
#> 1043           4         1.8     1
#> 1053           4         2.2     1
#> 14             5         0.2     2
#> 24             5         0.2     2
#> 34             5         0.2     2
#> 44             5         0.2     2
#> 59             5         0.2     2
#> 504            5         0.2     2
#> 514            5         1.4     1
#> 524            5         1.5     1
#> 534            5         1.5     1
#> 544            5         1.3     1
#> 554            5         1.5     1
#> 1004           5         1.3     1
#> 1014           5         2.5     1
#> 1024           5         1.9     1
#> 1034           5         2.1     1
#> 1044           5         1.8     1
#> 1054           5         2.2     1
#> 15             6         0.2     2
#> 25             6         0.2     2
#> 35             6         0.2     2
#> 45             6         0.2     2
#> 510            6         0.2     2
#> 505            6         0.2     2
#> 515            6         1.4     1
#> 525            6         1.5     1
#> 535            6         1.5     1
#> 545            6         1.3     1
#> 555            6         1.5     1
#> 1005           6         1.3     1
#> 1015           6         2.5     1
#> 1025           6         1.9     1
#> 1035           6         2.1     1
#> 1045           6         1.8     1
#> 1055           6         2.2     1
#> 16             7         0.2     2
#> 26             7         0.2     2
#> 36             7         0.2     2
#> 46             7         0.2     2
#> 516            7         0.2     2
#> 506            7         0.2     2
#> 517            7         1.4     1
#> 526            7         1.5     1
#> 536            7         1.5     1
#> 546            7         1.3     1
#> 556            7         1.5     1
#> 1006           7         1.3     1
#> 1016           7         2.5     1
#> 1026           7         1.9     1
#> 1036           7         2.1     1
#> 1046           7         1.8     1
#> 1056           7         2.2     1
#> 17             8         0.2     2
#> 27             8         0.2     2
#> 37             8         0.2     2
#> 47             8         0.2     2
#> 518            8         0.2     2
#> 507            8         0.2     2
#> 519            8         1.4     1
#> 527            8         1.5     1
#> 537            8         1.5     1
#> 547            8         1.3     1
#> 557            8         1.5     1
#> 1007           8         1.3     1
#> 1017           8         2.5     1
#> 1027           8         1.9     1
#> 1037           8         2.1     1
#> 1047           8         1.8     1
#> 1057           8         2.2     1
#> 18             9         0.2     2
#> 28             9         0.2     2
#> 38             9         0.2     2
#> 48             9         0.2     2
#> 520            9         0.2     2
#> 508            9         0.2     2
#> 5110           9         1.4     1
#> 528            9         1.5     1
#> 538            9         1.5     1
#> 548            9         1.3     1
#> 558            9         1.5     1
#> 1008           9         1.3     1
#> 1018           9         2.5     1
#> 1028           9         1.9     1
#> 1038           9         2.1     1
#> 1048           9         1.8     1
#> 1058           9         2.2     1
#> 19            10         0.2     2
#> 29            10         0.2     2
#> 39            10         0.2     2
#> 49            10         0.2     2
#> 529           10         0.2     2
#> 509           10         0.2     2
#> 5111          10         1.4     1
#> 5210          10         1.5     1
#> 539           10         1.5     1
#> 549           10         1.3     1
#> 559           10         1.5     1
#> 1009          10         1.3     1
#> 1019          10         2.5     1
#> 1029          10         1.9     1
#> 1039          10         2.1     1
#> 1049          10         1.8     1
#> 1059          10         2.2     1
#> 110           11         0.2     2
#> 210           11         0.2     2
#> 310           11         0.2     2
#> 410           11         0.2     2
#> 530           11         0.2     2
#> 5010          11         0.2     2
#> 5112          11         1.4     1
#> 5211          11         1.5     1
#> 5310          11         1.5     1
#> 5410          11         1.3     1
#> 5510          11         1.5     1
#> 10010         11         1.3     1
#> 10110         11         2.5     1
#> 10210         11         1.9     1
#> 10310         11         2.1     1
#> 10410         11         1.8     1
#> 10510         11         2.2     1
#> 111           12         0.2     2
#> 211           12         0.2     2
#> 311           12         0.2     2
#> 411           12         0.2     2
#> 540           12         0.2     2
#> 5011          12         0.2     2
#> 5113          12         1.4     1
#> 5212          12         1.5     1
#> 5311          12         1.5     1
#> 5411          12         1.3     1
#> 5511          12         1.5     1
#> 10011         12         1.3     1
#> 10111         12         2.5     1
#> 10211         12         1.9     1
#> 10311         12         2.1     1
#> 10411         12         1.8     1
#> 10511         12         2.2     1
#> 112           13         0.2     2
#> 212           13         0.2     2
#> 312           13         0.2     2
#> 412           13         0.2     2
#> 550           13         0.2     2
#> 5012          13         0.2     2
#> 5114          13         1.4     1
#> 5213          13         1.5     1
#> 5312          13         1.5     1
#> 5412          13         1.3     1
#> 5512          13         1.5     1
#> 10012         13         1.3     1
#> 10112         13         2.5     1
#> 10212         13         1.9     1
#> 10312         13         2.1     1
#> 10412         13         1.8     1
#> 10512         13         2.2     1
#> 113           14         0.2     2
#> 213           14         0.2     2
#> 313           14         0.2     2
#> 413           14         0.2     2
#> 560           14         0.2     2
#> 5013          14         0.2     2
#> 5115          14         1.4     1
#> 5214          14         1.5     1
#> 5313          14         1.5     1
#> 5413          14         1.3     1
#> 5513          14         1.5     1
#> 10013         14         1.3     1
#> 10113         14         2.5     1
#> 10213         14         1.9     1
#> 10313         14         2.1     1
#> 10413         14         1.8     1
#> 10513         14         2.2     1
#> 114           15         0.2     2
#> 214           15         0.2     2
#> 314           15         0.2     2
#> 414           15         0.2     2
#> 561           15         0.2     2
#> 5014          15         0.2     2
#> 5116          15         1.4     1
#> 5215          15         1.5     1
#> 5314          15         1.5     1
#> 5414          15         1.3     1
#> 5514          15         1.5     1
#> 10014         15         1.3     1
#> 10114         15         2.5     1
#> 10214         15         1.9     1
#> 10314         15         2.1     1
#> 10414         15         1.8     1
#> 10514         15         2.2     1
#> 115           16         0.2     2
#> 215           16         0.2     2
#> 315           16         0.2     2
#> 415           16         0.2     2
#> 562           16         0.2     2
#> 5015          16         0.2     2
#> 5117          16         1.4     1
#> 5216          16         1.5     1
#> 5315          16         1.5     1
#> 5415          16         1.3     1
#> 5515          16         1.5     1
#> 10015         16         1.3     1
#> 10115         16         2.5     1
#> 10215         16         1.9     1
#> 10315         16         2.1     1
#> 10415         16         1.8     1
#> 10515         16         2.2     1
#> 116           17         0.2     2
#> 216           17         0.2     2
#> 316           17         0.2     2
#> 416           17         0.2     2
#> 563           17         0.2     2
#> 5016          17         0.2     2
#> 5118          17         1.4     1
#> 5217          17         1.5     1
#> 5316          17         1.5     1
#> 5416          17         1.3     1
#> 5516          17         1.5     1
#> 10016         17         1.3     1
#> 10116         17         2.5     1
#> 10216         17         1.9     1
#> 10316         17         2.1     1
#> 10416         17         1.8     1
#> 10516         17         2.2     1
#> 117           18         0.2     2
#> 217           18         0.2     2
#> 317           18         0.2     2
#> 417           18         0.2     2
#> 564           18         0.2     2
#> 5017          18         0.2     2
#> 5119          18         1.4     1
#> 5218          18         1.5     1
#> 5317          18         1.5     1
#> 5417          18         1.3     1
#> 5517          18         1.5     1
#> 10017         18         1.3     1
#> 10117         18         2.5     1
#> 10217         18         1.9     1
#> 10317         18         2.1     1
#> 10417         18         1.8     1
#> 10517         18         2.2     1
#> 118           19         0.2     2
#> 218           19         0.2     2
#> 318           19         0.2     2
#> 418           19         0.2     2
#> 565           19         0.2     2
#> 5018          19         0.2     2
#> 5120          19         1.4     1
#> 5219          19         1.5     1
#> 5318          19         1.5     1
#> 5418          19         1.3     1
#> 5518          19         1.5     1
#> 10018         19         1.3     1
#> 10118         19         2.5     1
#> 10218         19         1.9     1
#> 10318         19         2.1     1
#> 10418         19         1.8     1
#> 10518         19         2.2     1
#> 119           20         0.2     2
#> 219           20         0.2     2
#> 319           20         0.2     2
#> 419           20         0.2     2
#> 566           20         0.2     2
#> 5019          20         0.2     2
#> 5121          20         1.4     1
#> 5220          20         1.5     1
#> 5319          20         1.5     1
#> 5419          20         1.3     1
#> 5519          20         1.5     1
#> 10019         20         1.3     1
#> 10119         20         2.5     1
#> 10219         20         1.9     1
#> 10319         20         2.1     1
#> 10419         20         1.8     1
#> 10519         20         2.2     1