Add significance asterisks to object

Takes an object, and adds significance asterisks.

est_sig(x, digits = 2, sig = NULL)

Arguments

x

An object for which a method exists. This will be treated as numeric by the default method.

digits

Integer. The number of digits to round the estimate column to.

sig

Optional, a vector of p-values for the default method.

Value

A character vector of formatted estimates.

See also

Author

Caspar J. van Lissa

Examples

est_sig(c(.222, .3333), sig = c(.054, .045))
#> [1] "0.22"  "0.33*"