Extract edges from a SEM model object

Attempts to extract edges from a SEM model object, where edges are defined as regression paths and covariances between variables (nodes).

get_edges(x, label = "est_sig", ...)

Arguments

x
A model object of class mplusObject or lavaan.

label
Either a character, indicating which column to use for edge labels, or an expression. See Details. Defaults to "est_sig", which labels edges with the estimated value with significance asterisks, as obtained from table_results. See Details and examples for more information.

...
Additional parameters passed to table_results. For example, users can pass the digits argument to control the number of digits in the edge label, or pass the columns argument to retain auxiliary columns in the tidy_edges data.frame for further processing (see Examples).

Value

An object of class 'tidy_edges'

Details

The function get_edges identifies all regression paths, latent variable definitions, and covariances in the model as edges. The output of table_results for those paths is used to label the edges.

## Custom labels

One way to create custom edge labels is by passing an expression to label. When an expression is passed to label, it is evaluated in the context of a data.frame containing the results of a call to table_results on the x argument.

Another way to create custom labels is by requesting auxiliary variables using the columns argument (which is passed to table_results), and then using these columns to construct a new label. See examples.

Examples

# Standard use
library(lavaan)
res <- sem("dist ~ speed", cars)
get_edges(res)

#>    from    to arrow     label connect_from connect_to curvature
#> 1 speed  dist  last   3.93***           NA         NA        NA
#> 2  dist  dist  both 227.07***           NA         NA        NA
#> 3 speed speed  both     27.40           NA         NA        NA


# Pass an expression to the 'label' argument for custom labels
get_edges(res, label = paste(est_sig, confint))

#>    from    to arrow                      label connect_from connect_to
#> 1 speed  dist  last       3.93*** [3.13, 4.73]           NA         NA
#> 2  dist  dist  both 227.07*** [138.06, 316.08]           NA         NA
#> 3 speed speed  both       27.40 [27.40, 27.40]           NA         NA
#>   curvature
#> 1        NA
#> 2        NA
#> 3        NA


# Pass the argument 'columns' to table_results through '...' to retain
# auxiliary columns for further processing
edg <- get_edges(res, columns = c("est_sig", "confint"))
edg

#>    from    to arrow     label   est_sig          confint connect_from
#> 1 speed  dist  last   3.93***   3.93***     [3.13, 4.73]           NA
#> 2  dist  dist  both 227.07*** 227.07*** [138.06, 316.08]           NA
#> 3 speed speed  both     27.40     27.40   [27.40, 27.40]           NA
#>   connect_to curvature
#> 1         NA        NA
#> 2         NA        NA
#> 3         NA        NA

edg <- within(edg, {label <- paste(est_sig, confint)})
edg

#>    from    to arrow                      label   est_sig          confint
#> 1 speed  dist  last       3.93*** [3.13, 4.73]   3.93***     [3.13, 4.73]
#> 2  dist  dist  both 227.07*** [138.06, 316.08] 227.07*** [138.06, 316.08]
#> 3 speed speed  both       27.40 [27.40, 27.40]     27.40   [27.40, 27.40]
#>   connect_from connect_to curvature
#> 1           NA         NA        NA
#> 2           NA         NA        NA
#> 3           NA         NA        NA