Apply pattern replacement over a vector

lsub returns a list of the same length as replacement, each element of which is the result of applying gsub to x using lapply.

lsub(x, replacement = NULL, pattern = "{C}", fixed = TRUE, ...)

Arguments

x

A character vector where matches are sought.

replacement

a character vector of length 1 or more. Each element is applied to x in turn. Default: NULL

pattern

A character string containing a regular expression (or character string when fixed = TRUE). Default: '{C}'.

fixed

logical. If TRUE, pattern is a string to be matched as is. Default: TRUE

...

Parameters passed on to gsub.

Value

A list of results returned by gsub.

Examples

lsub("a{C}", 1:3)
#> [[1]]
#> [1] "a1"
#> 
#> [[2]]
#> [1] "a2"
#> 
#> [[3]]
#> [1] "a3"
#>