Simulated depression data

This simulated dataset, based on work in progress by Plas and colleagues, contains six repeated measurements of the Depression subscale of the Symptom Checklist-90 (SCL-90).

Usage

data(plas_depression)

Format

A data frame with 978 rows and 6 variables.

Details

These data are inspired by the Prospection in Stress-related Military Research (PRISMO) study, which examined of psychological problems after deployment in more than 1,000 Dutch military personnel who were deployed to Afghanistan, from 2005-2019.

scl.1	integer	Sum score of SCL90 depression pre-deployment

scl.2	integer	Sum score of SCL90 depression 1 month post-deployment

scl.3	integer	Sum score of SCL90 depression 6 months post-deployment

scl.4	integer	Sum score of SCL90 depression 1 year post-deployment

scl.5	integer	Sum score of SCL90 depression 2 years post-deployment

scl.6	integer	Sum score of SCL90 depression 10 years post-deployment

References

van der Wal, S. J., Gorter, R., Reijnen, A., Geuze, E., & Vermetten, E. (2019). Cohort profile: The Prospective Research In Stress-Related Military Operations (PRISMO) study in the Dutch Armed Forces. BMJ Open, 9(3), e026670. doi:10.1136/bmjopen-2018-026670